Added typecast to floor function call to avoid error in VC.NET.
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@348 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/Array.cpp
@@ -1537,7 +1537,7 @@ break;
     Dimensions dim;
     int32 *rp = NULL;
     if (stepsize == 0) throw Exception("step size must be nonzero in colon expression");
-    int scount = (int) floor((maxval-minval)/stepsize) + 1;
+    int scount = (int) floor((float)((maxval-minval)/stepsize)) + 1;
     if (scount<=0) 
       dim.reset();
     else {
